#167 adding logout button
