+++ src/components/AccountItem/index.js
@@ -1,0 +1,2 @@
+import { AccountItem } from "./AccountItem";
+export default AccountItem;
